Move `FeedImageStore` to a separate file + separated `LocalFeedImageLoader` in extensions to better clarify its intents
