+++ EssentialFeed/Feed Cache/LocalFeedImageLoader.swift
@@ -1,22 +1,14 @@
 import Foundation
 
-public protocol FeedImageStore {
-    typealias RetrievalCompletion = (Result<Data, Error>) -> Void
-    typealias InsertCompletion = (Error?) -> Void
-
-    func retrieve(from url: URL, completion: @escaping RetrievalCompletion)
-    func insert(url: URL, with data: Data, completion: @escaping InsertCompletion)
-}
-
 public class LocalFeedImageLoader {
     private let store: FeedImageStore
-
-    public typealias LoadFeedImageResult = Result<Data, Error>
 
     public init(store: FeedImageStore) {
         self.store = store
     }
+}
 
+extension LocalFeedImageLoader: FeedImageLoader {
     private final class LocalFeedImageLoaderTask: FeedImageLoaderTask {
         private var completion: ((LoadFeedImageResult) -> Void)?
 
@@ -37,6 +29,8 @@ public class LocalFeedImageLoader {
         }
     }
 
+    public typealias LoadFeedImageResult = Result<Data, Error>
+
     public func load(from url: URL, completion: @escaping (LoadFeedImageResult) -> Void) -> FeedImageLoaderTask {
         let localTask = LocalFeedImageLoaderTask(completion)
 
@@ -55,7 +49,9 @@ public class LocalFeedImageLoader {
 
         return localTask
     }
+}
 
+extension LocalFeedImageLoader {
     public func save(url: URL, with data: Data, completion: @escaping (Error?) -> Void) {
         store.insert(url: url, with: data) { [weak self] error in
             guard self != nil else { return }
